# app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PhaseChangeDB API",
    "summary": "面向相变材料的证据化数据与发现 API。",
    "version": "0.2.0-mvp"
  },
  "paths": {
...
    "/v1/agent/query": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ask Scientist",
        "description": "Scientist Agent entry point.\n\nThe implementation must call domain tools/services. Direct SQL, direct\nElasticsearch queries, or direct Cypher from the LLM layer are forbidden.",
        "operationId": "ask_scientist_v1_agent_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/literature-parser/diagnostic": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "大模型连通性与配置安全诊断 (支持混合双模: 用户自备 Key / 服务端托管 Key)",
        "operationId": "diagnose_literature_agent_v1_agents_literature_parser_diagnostic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDiagnosticResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/literature-parser/extract": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "大模型相变物性结构化抽取并暂存至 ext_candidate（混合双模代理 + 领域 Schema 归一化）",
        "operationId": "extract_literature_properties_v1_agents_literature_parser_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/query": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ask Scientist",
        "description": "Scientist Agent entry point.\n\nThe implementation must call domain tools/services. Direct SQL, direct\nElasticsearch queries, or direct Cypher from the LLM layer are forbidden.",
        "operationId": "ask_scientist_api_v1_agent_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/literature-parser/diagnostic": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "大模型连通性与配置安全诊断 (支持混合双模: 用户自备 Key / 服务端托管 Key)",
        "operationId": "diagnose_literature_agent_api_v1_agents_literature_parser_diagnostic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDiagnosticResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/literature-parser/extract": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "大模型相变物性结构化抽取并暂存至 ext_candidate（混合双模代理 + 领域 Schema 归一化）",
        "operationId": "extract_literature_properties_api_v1_agents_literature_parser_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/AgentEvidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "tool_trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tool Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "confidence",
          "tool_trace_id"
        ],
        "title": "AgentAnswer"
      },
      "AgentAskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Question"
          },
          "material_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Material Ids"
          },
          "require_verified": {
            "type": "boolean",
            "title": "Require Verified",
            "default": true
          },
          "include_conflicting_evidence": {
            "type": "boolean",
            "title": "Include Conflicting Evidence",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AgentAskRequest"
      },
      "AgentDiagnosticRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "服务商: gemini 或 openai_compatible",
            "default": "gemini"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "自定义基准接口地址"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "模型标识符",
            "default": "gemini-2.5-pro"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "用户自定义 API Key (可选)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AgentDiagnosticRequest"
      },
      "AgentDiagnosticResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "诊断状态: success | warning | error"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "key_mode": {
            "type": "string",
            "title": "Key Mode",
            "description": "密钥模式: custom_user_key | server_hosted_key | demo_simulation"
          },
          "masked_key": {
            "type": "string",
            "title": "Masked Key"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "model_name",
          "provider",
          "key_mode",
          "masked_key",
          "latency_ms",
          "logs"
        ],
        "title": "AgentDiagnosticResponse"
      },
      "AgentEvidence": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "paper_id": {
            "type": "string",
            "format": "uuid",
            "title": "Paper Id"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "figure_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figure Number"
          },
          "table_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Number"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "stance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "paper_id"
        ],
        "title": "AgentEvidence"
      },
      "AgentExtractRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 10,
            "title": "Text",
            "description": "待抽取的文献全文、段落或图表上下文文本"
          },
          "paper_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Id",
            "description": "已入库文献的 UUID (可选)"
          },
          "paper_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Title",
            "description": "文献标题"
          },
          "paper_doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Doi",
            "description": "文献 DOI"
          },
          "first_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Author",
            "description": "第一作者"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "推理引擎提供商: gemini | openai_compatible",
            "default": "gemini"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "大模型标识",
            "default": "gemini-2.5-pro"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "用户自备 API Key (可选)"
          },
          "auto_stage": {
            "type": "boolean",
            "title": "Auto Stage",
            "description": "是否自动写入 ext_candidate 暂存区",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AgentExtractRequest"
      },
      "AgentExtractResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "key_mode": {
            "type": "string",
            "title": "Key Mode"
          },
          "candidates_extracted": {
            "type": "integer",
            "title": "Candidates Extracted"
          },
          "staged_candidate_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Staged Candidate Ids"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ExtractedPropertyItem"
            },
            "type": "array",
            "title": "Candidates"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_used",
          "provider",
          "key_mode",
          "candidates_extracted"
        ],
        "title": "AgentExtractResponse"
      },
...
      "ExtractedPropertyItem": {
        "properties": {
          "property_code": {
            "type": "string",
            "title": "Property Code"
          },
          "nominal_material": {
            "type": "string",
            "title": "Nominal Material"
          },
          "sample_formula": {
            "type": "string",
            "title": "Sample Formula"
          },
          "dopant_element": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dopant Element"
          },
          "dopant_ratio_at_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dopant Ratio At Pct"
          },
          "original_value": {
            "type": "number",
            "title": "Original Value"
          },
          "original_unit": {
            "type": "string",
            "title": "Original Unit"
          },
          "normalized_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Value"
          },
          "normalized_unit_term_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Unit Term Id"
          },
          "heating_rate_k_per_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heating Rate K Per Min"
          },
          "film_thickness_nm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Film Thickness Nm"
          },
          "substrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substrate"
          },
          "measurement_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Method"
          },
          "evidence_text": {
            "type": "string",
            "title": "Evidence Text"
          },
          "figure_or_table": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figure Or Table"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number",
            "default": 1
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.9
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "property_code",
          "nominal_material",
          "sample_formula",
          "original_value",
          "original_unit",
          "evidence_text"
        ],
        "title": "ExtractedPropertyItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}